# server/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Starliner API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/github/all-repositories/{organizationId}": {
            "get": {
                "tags": [
                    "github"
                ],
                "summary": "Get All Repositories (owner only, unfiltered)",
                "operationId": "getAllRepositories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Organization ID",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Repository"
                            }
                        }
                    }
                }
            }
        },
        "/github/repositories/{organizationId}": {
            "get": {
                "tags": [
                    "github"
                ],
                "summary": "Get Repositories",
                "operationId": "getRepositories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Organization ID",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Repository"
                            }
                        }
                    }
                }
            }
        },
        "/github/repositories/{organizationId}/{owner}/{repository}/contents": {
            "get": {
                "tags": [
                    "github"
                ],
                "summary": "Get Repository Content",
                "operationId": "getRepositoryContents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Organization ID",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository owner (user or org)",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "repository",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Path within the repository (e.g., src or src/main.go)",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.RepositoryFile"
                            }
                        }
                    }
                }
            }
        },
        "/github/repositories/{organizationId}/{owner}/{repository}/file": {
            "get": {
                "tags": [
                    "github"
                ],
                "summary": "Get File Content",
                "operationId": "getFileContent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Organization ID",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository owner (user or org)",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "repository",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Path to the file within the repository",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FileContent"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.FileContent": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
...
        "response.Repository": {
            "type": "object",
            "required": [
                "clone_url",
                "created_at",
                "full_name",
                "id",
                "name",
                "owner",
                "pushed_at",
                "updated_at"
            ],
            "properties": {
                "clone_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "pushed_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "response.RepositoryFile": {
            "type": "object",
            "required": [
                "name",
                "path",
                "sha",
                "size",
                "type",
                "url"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "sha": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}